pyproject.toml
[project]
name = "faculty-fetcher"
version = "0.1.0"
description = "This program scrapes the faculty list and saves them in a database"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "psycopg>=3.3.2",
    "psycopg-pool>=3.3.0",
    "pyqt6>=6.10.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "basedpyright>=1.37.1",
]
